\name{BBEST-package}
\alias{BBEST-package}
\alias{BBEST}
\docType{package}
\title{
Bayesian Background Estimation.
}
\description{
In this package we implemented a Bayesian-statistics approach for 
subtraction of incoherent scattering from neutron total-scattering data. 
In this approach, the estimated background signal associated with 
incoherent scattering maximizes the posterior probability, which combines 
the likelihood of this signal in reciprocal and real spaces with the prior 
that favors smooth lines.

To cite the BBEST package type: `citation("BBEST")' (without the single quotes).

For a listing of all routines in the BBEST package type: `library(help="BBEST")'

To start the Graphical User Interface type: `runUI()'

To start a simple command-line guide type: `guide()'
}
\details{
\tabular{ll}{
Package: \tab BBEST\cr
Type: \tab Package\cr
Version: \tab 0.1-0\cr
Date: \tab 2014-08-11\cr
License: \tab GPL-3\cr
}
}

\author{
Anton Gagin and Igor Levin with contributions from Charles R. Hogg III \cr
Maintainer: Anton Gagin <anton.gagin@nist.gov>
}
\references{
BBEST-package
}