\name{guide}
\alias{guide}

\title{
BBEST guide
}
\description{
\code{guide} is a function that guides through the Bayesian procedure for estimating the background 
}
\usage{
guide()
}

\value{ 
A list with elements:
\item{fit.res}{the return value of \code{\link{do.fit}}.}
\item{data}{an object of type \code{data}, see \code{\link{set.data}}.}
\item{gr}{the return value of \code{\link{calc.Gr}}.}
}

