\name{Internals}
\alias{Internals}
\alias{basisMatrix}
\alias{basisSpline}
\alias{DMatrix}
\alias{Dx}
\alias{bkg.analyt}
\alias{fplot.Gr}
\alias{get.bkg}
\alias{get.deriv}
\alias{get.deriv.numerically}
\alias{get.hess.numerically}
\alias{get.hess}
\alias{get.posterior}
\alias{golden.search}
\alias{grad.descent}
\alias{invert.order}
\alias{logLikelihoodBkg}
\alias{logLikelihoodBkgAnalyt}
\alias{logLikelihoodGrCorr}
\alias{logLikelihoodGrGauss}
\alias{logLikelihoodSignal}
\alias{logLikelihoodSignalAnalyt}
\alias{logPosterior}
\alias{logPosteriorAnalyt}
\alias{logPriorBkgRGP}
\alias{logPriorBkgRSmooth}
\alias{logPriorF}
\alias{logProbabilityBkgR}
\alias{noise.cov.matrix.r}
\alias{noise.cov.r}
\alias{noise.cov.vector.r}
\alias{regularized.cholesky}
\alias{sineFT.matrix}
\alias{sineFT}
\alias{step1}
\alias{step2}
\alias{step3}
\alias{step5}
\alias{step6}
\alias{step7}
\alias{step8}
\alias{step9}
\alias{step10}
\alias{set.DEoptim.control}
\alias{row.outer.product}
\alias{covMatrixSE}
\alias{covMatrix.DI}
\alias{whatIsSpecified}
\docType{data}
\alias{fpe}
\title{Functions used internally}
\description{Functions used internally}
\keyword{internal}

