\name{mPlot.sqa}
\alias{mPlot.sqa}
\title{
Plot the total normalized scattering intensity function \code{S(Q)} for individual detector banks
}
\description{
The function plots the total scattering functions \code{S(Q)} returned by \code{PDFgetN} in \code{\link{read.sqa}}.
}
\usage{
mPlot.sqa(data)
}
\arguments{
  \item{data}{list, the return value of \code{\link{read.sqa}}.}
}

\seealso{\code{\link{read.sqa}}}

\keyword{plot}
\keyword{PDFgetN}
