\name{sqa.split}
\alias{sqa.split}

\title{
Split .sqa file into individual files for each databank 
}
\description{
\code{sqa.split} splits \emph{PDFgetN} .\emph{sqa}-file into individual files for each databank. 
}
\usage{
sqa.split(file = stop("'file' must be specified"))
}
\arguments{
  \item{file}{character, name of the source file.}
}

\seealso{
\code{\link{read.sqa}}, \code{\link{do.fit.banks}}, \code{\link{BBEST-package}}
}

\keyword{PDFgetN}
\keyword{banks}
\keyword{save data}