\name{runUI}
\alias{runUI}
\title{Start the GUI}
\description{
Starts the application and opens up the default web browser to view it.
}
\usage{
runUI()
}
\details{
Runs a \bold{Shiny} application. This function normally does not return; interrupt bold{R} to stop the application (usually by pressing Ctrl+C or Esc).
}

\keyword{GUI}
